Fix make test-all + Feedback changes
