+++ circuit-benchmarks/src/super_circuit.rs
@@ -1,378 +1,7 @@
 //! SuperCircuit circuit benchmarks
 
-use eth_types::{Bytes, U256};
-use halo2_proofs::poly::commitment::Params;
-use serde::{Deserialize, Serialize};
-use std::{fs, rc::Rc};
-use zkevm_circuits::{
-    root_circuit::{
-        taiko_aggregation::AccumulationSchemeType, KzgDk, KzgSvk, TaikoAggregationCircuit,
-    },
-    taiko_super_circuit::{test::block_1tx, SuperCircuit},
-    witness::ProtocolInstance,
-};
-
-use bus_mapping::circuit_input_builder::CircuitsParams;
-
-use rand::SeedableRng;
-
-use ark_std::{end_timer, start_timer};
-use halo2_proofs::{
-    halo2curves::{
-        bn256::{Bn256, Fq, Fr, G1Affine},
-        ff::PrimeField,
-    },
-    plonk::{keygen_pk, keygen_vk, ProvingKey, VerifyingKey},
-    poly::{
-        commitment::ParamsProver,
-        kzg::commitment::{KZGCommitmentScheme, ParamsKZG},
-    },
-};
-use std::path::Path;
-
-use snark_verifier_sdk::{
-    evm::{gen_evm_proof_gwc, gen_evm_proof_shplonk},
-    halo2::{aggregation::AccumulationSchemeSDK, gen_snark_gwc, gen_snark_shplonk, gen_srs},
-    Snark, GWC, SHPLONK,
-};
-
-use snark_verifier::{
-    loader::evm::{self, encode_calldata, Address as VerifierAddress, EvmLoader, ExecutorBuilder},
-    pcs::kzg::*,
-    system::halo2::{compile, transcript::evm::EvmTranscript, Config},
-    verifier::SnarkVerifier,
-};
-
-/// Number of limbs to decompose a elliptic curve base field element into.
-pub const LIMBS: usize = 4;
-/// Number of bits of each decomposed limb.
-pub const BITS: usize = 68;
-
-pub type ProverParams = ParamsKZG<Bn256>;
-pub type ProverCommitmentScheme = KZGCommitmentScheme<Bn256>;
-pub type ProverKey = ProvingKey<G1Affine>;
-
-/// KZG accumulation scheme with GWC19 multiopen.
-pub type PlonkVerifierGWC =
-    snark_verifier::verifier::plonk::PlonkVerifier<GWC, LimbsEncoding<LIMBS, BITS>>;
-
-pub type PlonkVerifierSHPLONK =
-    snark_verifier::verifier::plonk::PlonkVerifier<SHPLONK, LimbsEncoding<LIMBS, BITS>>;
-
-use rand::rngs::StdRng;
-
-#[derive(Serialize, Deserialize, Debug)]
-struct BlockProofData {
-    instances: Vec<U256>,
-    proof: Bytes,
-}
-
-/// Fixed rng for testing purposes
-pub fn fixed_rng() -> StdRng {
-    StdRng::seed_from_u64(9)
-}
-
-/// Returns [<len>, ...] of `instance`
-pub fn gen_num_instance(instance: &[Vec<Fr>]) -> Vec<usize> {
-    instance.iter().map(|v| v.len()).collect()
-}
-
-fn gen_verifier(
-    params: &ProverParams,
-    vk: &VerifyingKey<G1Affine>,
-    config: Config,
-    num_instance: Vec<usize>,
-    aggregation_type: AccumulationSchemeType,
-) -> String {
-    let protocol = compile(params, vk, config);
-    let svk = KzgSvk::<Bn256>::new(params.get_g()[0]);
-    let dk = KzgDk::<Bn256>::new(svk, params.g2(), params.s_g2());
-
-    let loader = EvmLoader::new::<Fq, Fr>();
-    let protocol = protocol.loaded(&loader);
-    let mut transcript = EvmTranscript::<_, Rc<EvmLoader>, _, _>::new(&loader);
-
-    let instances = transcript.load_instances(num_instance);
-    match aggregation_type {
-        AccumulationSchemeType::GwcType => {
-            let proof =
-                PlonkVerifierGWC::read_proof(&dk, &protocol, &instances, &mut transcript).unwrap();
-            PlonkVerifierGWC::verify(&dk, &protocol, &instances, &proof).unwrap();
-        }
-        AccumulationSchemeType::ShplonkType => {
-            let proof =
-                PlonkVerifierSHPLONK::read_proof(&dk, &protocol, &instances, &mut transcript)
-                    .unwrap();
-            PlonkVerifierSHPLONK::verify(&dk, &protocol, &instances, &proof).unwrap();
-        }
-    };
-
-    let yul = loader.yul_code();
-    fs::write(Path::new("./aggregation_plonk.yul"), &yul).unwrap();
-    yul
-}
-
-fn evm_verify(deployment_code: Vec<u8>, instances: Vec<Vec<Fr>>, proof: Vec<u8>) {
-    let calldata = encode_calldata(&instances, &proof);
-    println!(
-        "deploy code size: {} bytes, instances size: [{}][{}], calldata: {}",
-        deployment_code.len(),
-        instances.len(),
-        instances[0].len(),
-        calldata.len(),
-    );
-
-    let success = {
-        let mut evm = ExecutorBuilder::default()
-            .with_gas_limit(u64::MAX.into())
-            .build();
-        let caller = VerifierAddress::from_low_u64_be(0xfe);
-        let deploy = evm.deploy(caller, deployment_code.into(), 0.into());
-        match deploy.address {
-            Some(addr) => {
-                let result = evm.call_raw(caller, addr, calldata.into(), 0.into());
-                println!("Deployment cost: {} gas", deploy.gas_used);
-                println!("Verification cost: {} gas", result.gas_used);
-                !result.reverted
-            }
-            None => {
-                println!("Deployment failed due to {:?}", deploy.exit_reason);
-                false
-            }
-        }
-    };
-    assert!(success);
-}
-
-fn gen_application_snark(
-    params: &ParamsKZG<Bn256>,
-    aggregation_type: AccumulationSchemeType,
-) -> Snark {
-    println!("gen app snark");
-    let circuits_params = CircuitsParams {
-        max_txs: 2,
-        max_calldata: 200,
-        max_rws: 256,
-        max_copy_rows: 256,
-        max_exp_steps: 256,
-        max_bytecode: 512,
-        max_evm_rows: 0,
-        max_keccak_rows: 0,
-    };
-    let protocol_instance = ProtocolInstance {
-        anchor_gas_limit: 150000,
-        ..Default::default()
-    };
-    let (_, super_circuit, _, _) = SuperCircuit::<_>::build(
-        block_1tx(&protocol_instance),
-        circuits_params,
-        protocol_instance,
-    )
-    .unwrap();
-
-    // let pk = gen_pk(params, &super_circuit, Some(Path::new("./examples/app.pk")),
-    // super_circuit.params());
-    let vk = keygen_vk(params, &super_circuit).expect("keygen_vk should not fail");
-    let pk = keygen_pk(params, vk, &super_circuit).expect("keygen_pk should not fail");
-    match aggregation_type {
-        AccumulationSchemeType::GwcType => gen_snark_gwc(params, &pk, super_circuit, None::<&str>),
-        AccumulationSchemeType::ShplonkType => {
-            gen_snark_shplonk(params, &pk, super_circuit, None::<&str>)
-        }
-    }
-}
-
-fn create_root_super_circuit_prover_sdk<const T: u64, AS: AccumulationSchemeSDK>() {
-    let params_app = gen_srs(18);
-    let aggregation_type = T.into();
-    let snarks = [(); 1].map(|_| gen_application_snark(&params_app, aggregation_type));
-
-    let params = gen_srs(22);
-    let mut snark_roots = Vec::new();
-    for snark in snarks {
-        let pcd_circuit = TaikoAggregationCircuit::<AS>::new(&params, [snark]).unwrap();
-
-        let start0 = start_timer!(|| "gen vk & pk");
-        // let pk = gen_pk(
-        // &params,
-        // &agg_circuit.without_witnesses(),
-        // Some(Path::new("./examples/agg.pk")),
-        // agg_circuit.params(),
-        // );
-        let vk = keygen_vk(&params, &pcd_circuit).expect("keygen_vk should not fail");
-        let pk = keygen_pk(&params, vk, &pcd_circuit).expect("keygen_pk should not fail");
-        end_timer!(start0);
-
-        let _root = match aggregation_type {
-            AccumulationSchemeType::GwcType => {
-                gen_snark_gwc(
-                    &params,
-                    &pk,
-                    pcd_circuit.clone(),
-                    // Some(Path::new("./examples/agg.snark"))
-                    None::<&str>,
-                )
-            }
-            AccumulationSchemeType::ShplonkType => {
-                gen_snark_shplonk(
-                    &params,
-                    &pk,
-                    pcd_circuit.clone(),
-                    // Some(Path::new("./examples/agg.snark"))
-                    None::<&str>,
-                )
-            }
-        };
-
-        snark_roots.push(_root);
-    }
-
-    println!("gen blocks agg snark");
-    let params = gen_srs(22);
-    let agg_circuit = TaikoAggregationCircuit::<AS>::new(&params, snark_roots).unwrap();
-    println!("new root agg circuit {}", agg_circuit);
-
-    let start0 = start_timer!(|| "gen vk & pk");
-    // let pk = gen_pk(
-    // &params,
-    // &agg_circuit.without_witnesses(),
-    // Some(Path::new("./examples/agg.pk")),
-    // agg_circuit.params(),
-    // );
-    let vk = keygen_vk(&params, &agg_circuit).expect("keygen_vk should not fail");
-    let pk = keygen_pk(&params, vk.clone(), &agg_circuit).expect("keygen_pk should not fail");
-    end_timer!(start0);
-
-    // std::fs::remove_file("./examples/agg.snark").unwrap_or_default();
-    // let _snark = gen_snark_shplonk(
-    // &params,
-    // &pk,
-    // agg_circuit.clone(),
-    // Some(Path::new("./examples/agg.snark"))*/None::<&str>,
-    // );
-
-    println!("gen evm snark");
-    // do one more time to verify
-    let num_instances = agg_circuit.num_instance();
-    let instances = agg_circuit.instance();
-    let accumulator_indices = Some(agg_circuit.accumulator_indices());
-    let proof_calldata = match aggregation_type {
-        AccumulationSchemeType::GwcType => {
-            gen_evm_proof_gwc(&params, &pk, agg_circuit, instances.clone())
-        }
-        AccumulationSchemeType::ShplonkType => {
-            gen_evm_proof_shplonk(&params, &pk, agg_circuit, instances.clone())
-        }
-    };
-
-    let block_proof_data = BlockProofData {
-        instances: instances
-            .iter()
-            .flatten()
-            .map(|v| U256::from_little_endian(v.to_repr().as_ref()))
-            .collect(),
-        proof: proof_calldata.clone().into(),
-    };
-
-    fs::write(
-        Path::new("./proof.json"),
-        serde_json::to_vec(&block_proof_data).unwrap(),
-    )
-    .unwrap();
-
-    let deployment_code = gen_verifier(
-        &params,
-        &vk,
-        Config::kzg()
-            .with_num_instance(num_instances.clone())
-            .with_accumulator_indices(accumulator_indices),
-        num_instances,
-        aggregation_type,
-    );
-    let evm_verifier_bytecode = evm::compile_yul(&deployment_code);
-
-    evm_verify(evm_verifier_bytecode, instances, proof_calldata);
-}
-
-// for N super circuit -> 1 root circuit integration
-fn create_1_level_root_super_circuit_prover_sdk<const T: u64, AS: AccumulationSchemeSDK>() {
-    let agg_type = T.into();
-    let app_degree = 18;
-    let min_k_aggretation = 22;
-    let mut params_app = gen_srs(min_k_aggretation);
-    params_app.downsize(app_degree);
-    let snarks = [(); 1].map(|_| gen_application_snark(&params_app, agg_type));
-
-    println!("gen blocks agg snark");
-    let params = gen_srs(min_k_aggretation);
-    let agg_circuit = TaikoAggregationCircuit::<AS>::new(&params, snarks).unwrap();
-    let start0 = start_timer!(|| "gen vk & pk");
-    // let pk = gen_pk(
-    // &params,
-    // &agg_circuit.without_witnesses(),
-    // Some(Path::new("./examples/agg.pk")),
-    // agg_circuit.params(),
-    // );
-    let vk = keygen_vk(&params, &agg_circuit).expect("keygen_vk should not fail");
-    let pk = keygen_pk(&params, vk.clone(), &agg_circuit).expect("keygen_pk should not fail");
-    end_timer!(start0);
-
-    // std::fs::remove_file("./examples/agg.snark").unwrap_or_default();
-    // let _snark = gen_snark_shplonk(
-    // &params,
-    // &pk,
-    // agg_circuit.clone(),
-    // Some(Path::new("./examples/agg.snark"))*/None::<&str>,
-    // );
-
-    println!("gen evm snark");
-    // do one more time to verify
-    let num_instances = agg_circuit.num_instance();
-    let instances = agg_circuit.instance();
-    let accumulator_indices = Some(agg_circuit.accumulator_indices());
-    let proof_calldata = match agg_type {
-        AccumulationSchemeType::GwcType => {
-            gen_evm_proof_gwc(&params, &pk, agg_circuit, instances.clone())
-        }
-        AccumulationSchemeType::ShplonkType => {
-            gen_evm_proof_shplonk(&params, &pk, agg_circuit, instances.clone())
-        }
-    };
-
-    let block_proof_data = BlockProofData {
-        instances: instances
-            .iter()
-            .flatten()
-            .map(|v| U256::from_little_endian(v.to_repr().as_ref()))
-            .collect(),
-        proof: proof_calldata.clone().into(),
-    };
-
-    fs::write(
-        Path::new("./proof.json"),
-        serde_json::to_vec(&block_proof_data).unwrap(),
-    )
-    .unwrap();
-
-    let deployment_code = gen_verifier(
-        &params,
-        &vk,
-        Config::kzg()
-            .with_num_instance(num_instances.clone())
-            .with_accumulator_indices(accumulator_indices),
-        num_instances,
-        agg_type,
-    );
-    let evm_verifier_bytecode = evm::compile_yul(&deployment_code);
-
-    evm_verify(evm_verifier_bytecode, instances, proof_calldata);
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::super_circuit::{
-        create_1_level_root_super_circuit_prover_sdk, create_root_super_circuit_prover_sdk,
-    };
     use ark_std::{end_timer, start_timer};
     use bus_mapping::circuit_input_builder::CircuitsParams;
     use eth_types::{address, bytecode, geth_types::GethData, Word};
@@ -394,34 +23,9 @@ mod tests {
     };
     use mock::{TestContext, MOCK_CHAIN_ID};
     use rand::SeedableRng;
-    use rand_chacha::{ChaCha20Rng, ChaChaRng};
-    use snark_verifier_sdk::{halo2::read_or_create_srs, GWC, SHPLONK};
+    use rand_chacha::ChaChaRng;
     use std::{collections::HashMap, env::var};
-    use zkevm_circuits::{
-        root_circuit::taiko_aggregation::AccumulationSchemeType, super_circuit::SuperCircuit,
-    };
-
-    #[test]
-    fn test_setup_random_params() {
-        let k = 22;
-        read_or_create_srs::<G1Affine, _>(k, move |k| {
-            ParamsKZG::<Bn256>::setup(k, ChaCha20Rng::from_entropy())
-        });
-    }
-
-    #[test]
-    fn bench_root_super_circuit_prover_sdk() {
-        // New version, cleanest using the new sdk
-        const AGG_TYPE: u64 = AccumulationSchemeType::GwcType as u64;
-        create_root_super_circuit_prover_sdk::<AGG_TYPE, GWC>();
-    }
-
-    #[test]
-    fn bench_n_to_1_root_super_circuit_prover() {
-        // for N->1 aggregation using new sdk
-        const AGG_TYPE: u64 = AccumulationSchemeType::ShplonkType as u64;
-        create_1_level_root_super_circuit_prover_sdk::<AGG_TYPE, SHPLONK>();
-    }
+    use zkevm_circuits::super_circuit::SuperCircuit;
 
     #[cfg_attr(not(feature = "benches"), ignore)]
     #[test]
