+++ zkevm-circuits/src/lib.rs
@@ -23,11 +23,11 @@ pub mod evm_circuit;
 pub mod exp_circuit;
 pub mod keccak_circuit;
 pub mod pi_circuit;
-pub mod pi_circuit2;
 pub mod root_circuit;
 pub mod state_circuit;
 pub mod super_circuit;
 pub mod table;
+pub mod taiko_pi_circuit;
 pub mod taiko_super_circuit;
 
 #[cfg(any(feature = "test", test))]
